requirements.txt
discord.py==1.3.3
asyncio==3.4.3
mysql-connector==2.2.9
pytz==2020.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.12.0
